+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 22.2
++ Add optionally visible comments in `rvx/dom`.
++ Rename `NoopComment` to `Comment`.
++ **Deprecated:** `NoopComment` export will be removed in the next major version.
+
 ## 22.1
 + Add `separated` conversion utility.